Attempt at pointing EP0_RxReady to proper classes (scuffed)

+++ Core/Src/main.c
@@ -58,7 +58,7 @@ static void MX_GPIO_Init(void);
 /* Private user code ---------------------------------------------------------*/
 /* USER CODE BEGIN 0 */
 int _write(int file, char *ptr, int len) {
-	CDC_Transmit_FS(ptr, len);
+	CDC_Transmit_FS(ptr, (uint16_t)len);
     return len;
 }
 /* USER CODE END 0 */
@@ -125,6 +125,7 @@ int main(void)
 		  lcd_clear();
 		  lcd_put_str_at("Button 1", 13, 1, 0);
 		  printf("Button 1 pressed\r\n");
+		  //printf("test");
 	  }
 	  if(HAL_GPIO_ReadPin (BTN_2_GPIO_Port, BTN_2_Pin)) { // WARNING: This logic is faulty in the USB report cause of HAL_Delay on low clock speed (getting errors when delay is only at 10ms)
 		  //printf("Button 2 pressed\r\n");
